requirements.txt
alembic==1.17.2
aniso8601==10.0.1
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
